requirements.txt
streamlit>=1.30.0
litellm>=1.34.0
asynciolimiter>=1.0.0
aiohttp>=3.9.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
parallel>=0.1.0
